Develop

+++ src/controllers/cards.ts
@@ -1,78 +1,58 @@
-import { Request, Response } from 'express';
+import { Request, Response, NextFunction } from 'express';
 import { ObjectId } from 'mongoose';
 import card from '../models/card';
-import {DEFAULT_ERROR, NOT_FOUND, WRONG_DATA} from '../utils/response-errors';
+import { IExtendedRequestId } from '../types/model-types';
 
-export const getCards = (req: Request, res: Response) => {
+export const getCards = (req: Request, res: Response, next: NextFunction) => {
   card.find({})
     .then((info) => res.send({ info }))
-    .catch(() => res.status(DEFAULT_ERROR).send({ message: 'Ошибка сервера' }));
+    .catch(next);
 };
 
-export const createCard = (req: Request, res: Response) => {
-  const id = req.user?._id;
+export const createCard = (req: IExtendedRequestId, res: Response, next: NextFunction) => {
+  const id = req.user && req.user._id;
   const { name, link } = req.body;
   card.create({ name, link, owner: id })
     .then((info) => res.status(201).send({ info }))
-    .catch((err) => {
-      if (err.name === 'ValidationError') {
-        res.status(WRONG_DATA).send({ message: 'Переданы некорректные данные для создания карточки' });
-      } else {
-        res.status(DEFAULT_ERROR).send({ message: 'Ошибка сервера' });
-      }
-    });
+    .catch(next);
 };
 
-export const deleteCard = (req: Request, res: Response) => {
+export const deleteCard = (req: IExtendedRequestId, res: Response, next: NextFunction) => {
   const { cardId } = req.params;
-  card.deleteOne({ cardId })
-    .orFail(new Error('NotFound'))
-    .then((info) => res.send({ info }))
-    .catch((err) => {
-      if (err.name === 'CastError') {
-        res.status(WRONG_DATA).send({ message: 'Переданы некорректные данные' });
-      } else if (err.message === 'NotFound') {
-        res.status(NOT_FOUND).send({ message: 'Объект не найден' });
-      } else {
-        res.status(DEFAULT_ERROR).send({ message: 'Ошибка сервера' });
+  const id = req.user && req.user._id;
+  card.findOne({ cardId })
+    // .orFail(new Error('NotFound'))
+    .then((info) => {
+      if (!info) {
+        throw new Error('Карточка не найдена');
       }
-    });
+      if (info.owner.toString() !== id) {
+        throw new Error('Карточка другого пользователя');
+      }
+      return card.deleteOne({ cardId });
+    })
+    .then((info) => res.send({ info }))
+    .catch(next);
 };
 
-export const likeCard = (req: Request, res: Response) => {
+export const likeCard = (req: IExtendedRequestId, res: Response, next: NextFunction) => {
   card.findByIdAndUpdate(
     req.params.cardId,
-    { $addToSet: { likes: req.user._id } },
+    { $addToSet: { likes: req.user && req.user._id } },
     { new: true },
   )
     .orFail(new Error('NotFound'))
     .then((info) => res.send({ info }))
-    .catch((err) => {
-      if (err.name === 'CastError') {
-        res.status(WRONG_DATA).send({ message: 'Переданы некорректные данные' });
-      } else if (err.message === 'NotFound') {
-        res.status(NOT_FOUND).send({ message: 'Объект не найден' });
-      } else {
-        res.status(DEFAULT_ERROR).send({ message: 'Ошибка сервера' });
-      }
-    });
+    .catch(next);
 };
 
-export const dislikeCard = (req: Request, res: Response) => {
+export const dislikeCard = (req: IExtendedRequestId, res: Response, next: NextFunction) => {
   card.findByIdAndUpdate(
     req.params.cardId,
-    { $pull: { likes: req.user._id as ObjectId } },
+    { $pull: { likes: req.user && req.user._id as ObjectId } },
     { new: true },
   )
     .orFail(new Error('NotFound'))
     .then((info) => res.send({ info }))
-    .catch((err) => {
-      if (err.name === 'CastError') {
-        res.status(WRONG_DATA).send({ message: 'Переданы некорректные данные' });
-      } else if (err.message === 'NotFound') {
-        res.status(NOT_FOUND).send({ message: 'Объект не найден' });
-      } else {
-        res.status(DEFAULT_ERROR).send({ message: 'Ошибка сервера' });
-      }
-    });
+    .catch(next);
 };
